fix: "btw": trailing space, not tab
Fixes the "simplest example" of btw -> by the way so that it doesn't
result in a trailing tab, which is caused by tab being the completion
character.
Using the leading `:o` stops that.
And then we end with a trailing space and backtick to output a space at
the end, after the "way".

This fixes #2, "fix btw to end with a space not a tab".