feat: Convert `etc` to "and so on."
For the moment, for convenience, we end with ". ", as it's usually the
end of a thought. But modify later, if it proves inconvenient.

Not only is it part of our general approach to plain English, it's also
mentioned by the
[UK Gov style
guide](https://insidegovuk.blog.gov.uk/2016/07/20/changes-to-the-style-guide-no-more-eg-and-ie-etc/).

> ’etc’ can usually be avoided. Try using ‘for example’ or ‘such as’ or
> ‘including’. Never use ‘etc’ at the end of a list starting with ‘for
> example’ or ‘such as’ or ‘including’.

FIxes #4, 'Convert etc to "and so on"'